+++ examples/basic/styleguide.config.js
@@ -17,6 +17,7 @@ const transpileDependencies = [
 
 const docSiteUrl = process.env.DEPLOY_PRIME_URL || 'https://vue-styleguidist.github.io'
 
+/** @type import("vue-styleguidist").Config */
 module.exports = {
 	title: 'Vue Styleguidist basic',
 	components: 'src/components/**/[A-Z]*.vue',