+++ packages/vue-docgen-api/src/script-handlers/__tests__/propHandler.ts
@@ -1,4 +1,5 @@
 import { ParserPlugin } from '@babel/parser'
+import * as bt from '@babel/types'
 import { NodePath } from 'ast-types/lib/node-path'
 import babylon from '../../babel-parser'
 import Documentation, { PropDescriptor } from '../../Documentation'
@@ -23,6 +24,12 @@ describe('propHandler', () => {
 	let documentation: Documentation
 	let mockPropDescriptor: PropDescriptor
 
+	let defaultAST: bt.File
+	const options = { filePath: '', validExtends: () => true }
+	beforeAll(() => {
+		defaultAST = babylon({ plugins: ['typescript'] }).parse('const a  = 1')
+	})
+
 	beforeEach(() => {
 		mockPropDescriptor = {
 			description: '',
@@ -35,23 +42,27 @@ describe('propHandler', () => {
 		mockGetPropDescriptor.mockReturnValue(mockPropDescriptor)
 	})
 
-	function parserTest(src: string, plugins?: ParserPlugin[]): PropDescriptor {
+	async function parserTest(
+		src: string,
+		plugins?: ParserPlugin[],
+		ast = defaultAST
+	): Promise<PropDescriptor> {
 		const def = parse(src, plugins)
 		if (def) {
-			propHandler(documentation, def)
+			await propHandler(documentation, def, ast, options)
 		}
 		return mockPropDescriptor
 	}
 
 	describe('base', () => {
-		it('should accept an array of string as props', () => {
+		it('should accept an array of string as props', async () => {
 			const src = `
         export default {
           props: ['testArray']
         }`
 			const def = parse(src)
 			if (def) {
-				propHandler(documentation, def)
+				await propHandler(documentation, def, defaultAST, options)
 			}
 			expect(mockPropDescriptor.required).toBeFalsy()
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('testArray')
@@ -59,7 +70,7 @@ describe('propHandler', () => {
 	})
 
 	describe('type', () => {
-		it('should return the right props type', () => {
+		it('should return the right props type', async () => {
 			const src = `
         export default {
           name: 'name-123',
@@ -73,12 +84,12 @@ describe('propHandler', () => {
           }
         }
 		`
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'array' }
 			})
 		})
 
-		it('should return the right props type for lit Array', () => {
+		it('should return the right props type for lit Array', async () => {
 			const src = `
         export default {
           props: {
@@ -86,12 +97,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'array' }
 			})
 		})
 
-		it('should return the right props composite type', () => {
+		it('should return the right props composite type', async () => {
 			const src = `
         export default {
           props: {
@@ -101,12 +112,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'string|number' }
 			})
 		})
 
-		it('should return the right props type for Array', () => {
+		it('should return the right props type for Array', async () => {
 			const src = `
         export default {
           props: {
@@ -114,12 +125,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'array' }
 			})
 		})
 
-		it('should not return required if prop only Type', () => {
+		it('should not return required if prop only Type', async () => {
 			const src = `
         export default {
           props: {
@@ -129,12 +140,12 @@ describe('propHandler', () => {
         `
 			const def = parse(src)
 			if (def) {
-				propHandler(documentation, def)
+				propHandler(documentation, def, defaultAST, options)
 			}
 			expect(mockPropDescriptor.required).toBeUndefined()
 		})
 
-		it('should return the right props type string', () => {
+		it('should return the right props type string', async () => {
 			const src = `
         export default {
           props: {
@@ -142,12 +153,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'string' }
 			})
 		})
 
-		it('should return the right props composite string|number', () => {
+		it('should return the right props composite string|number', async () => {
 			const src = `
         export default {
           props: {
@@ -155,12 +166,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'string|number' }
 			})
 		})
 
-		it('should deduce the prop type from the default value', () => {
+		it('should deduce the prop type from the default value', async () => {
 			const src = `
         export default {
           props: {
@@ -170,12 +181,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: { name: 'boolean' }
 			})
 		})
 
-		it('should still return props with vue-types', () => {
+		it('should still return props with vue-types', async () => {
 			const src = [
 				'export default {',
 				'  props:{',
@@ -186,14 +197,14 @@ describe('propHandler', () => {
 				'  }',
 				'}'
 			].join('\n')
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: {
 					func: true
 				}
 			})
 		})
 
-		it('should still return props with prop-types', () => {
+		it('should still return props with prop-types', async () => {
 			const src = [
 				'export default {',
 				'  props:{',
@@ -201,23 +212,23 @@ describe('propHandler', () => {
 				'  }',
 				'}'
 			].join('\n')
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: {
 					func: true
 				}
 			})
 		})
 
-		it('should still return props with delegated types', () => {
+		it('should still return props with delegated types', async () => {
 			const src = ['export default {', '  props: {', '    toto', '  }', '}'].join('\n')
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: {}
 			})
 		})
 	})
 
 	describe('required', () => {
-		it('should return the right required props', () => {
+		it('should return the right required props', async () => {
 			const src = `
         export default {
           name: 'name-123',
@@ -231,14 +242,14 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				required: true
 			})
 		})
 	})
 
 	describe('defaultValue', () => {
-		it('should be ok with just the default', () => {
+		it('should be ok with just the default', async () => {
 			const src = `
         export default {
           props: {
@@ -248,12 +259,12 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				defaultValue: { value: `"normal"` }
 			})
 		})
 
-		it('should be ok with the default as a method', () => {
+		it('should be ok with the default as a method', async () => {
 			const src = [
 				'export default {',
 				'  props: {',
@@ -266,7 +277,7 @@ describe('propHandler', () => {
 				'}'
 			].join('\n')
 
-			expect(parserTest(src).defaultValue).toMatchInlineSnapshot(`
+			expect((await parserTest(src)).defaultValue).toMatchInlineSnapshot(`
       Object {
         "func": true,
         "value": "function() {
@@ -276,7 +287,7 @@ describe('propHandler', () => {
     `)
 		})
 
-		it('should deal properly with multilple returns', () => {
+		it('should deal properly with multilple returns', async () => {
 			const src = `
         export default {
           props: {
@@ -293,7 +304,7 @@ describe('propHandler', () => {
           }
         }
         `
-			const testParsed = parserTest(src)
+			const testParsed = await parserTest(src)
 			const defaultValue = removeWhitespaceForTest(testParsed.defaultValue)
 			expect(defaultValue).toMatchObject({
 				func: true,
@@ -301,7 +312,7 @@ describe('propHandler', () => {
 			})
 		})
 
-		it('should deal properly with multilple returns in arrow functions', () => {
+		it('should deal properly with multilple returns in arrow functions', async () => {
 			const src = `
         export default {
           props: {
@@ -318,7 +329,7 @@ describe('propHandler', () => {
           }
         }
         `
-			const testParsed = parserTest(src)
+			const testParsed = await parserTest(src)
 			const defaultValue = removeWhitespaceForTest(testParsed.defaultValue)
 			expect(defaultValue).toMatchObject({
 				func: true,
@@ -353,7 +364,7 @@ describe('propHandler', () => {
 			]
 		])(
 			'if prop is of type %p,\n\t given %p as default,\n\t should parse as %p,\n\t comment types are %p',
-			(propType, input, output, commentsBlockType) => {
+			async (propType, input, output, commentsBlockType) => {
 				const src = `
                 export default {
                   props: {
@@ -367,7 +378,7 @@ describe('propHandler', () => {
                   }
                 }
                 `
-				const testParsed = parserTest(src)
+				const testParsed = await parserTest(src)
 				const defaultValue = removeWhitespaceForTest(testParsed.defaultValue)
 				expect(defaultValue).toMatchObject({ value: output })
 			}
@@ -375,7 +386,7 @@ describe('propHandler', () => {
 	})
 
 	describe('description', () => {
-		it('should return the right description', () => {
+		it('should return the right description', async () => {
 			const src = `
         export default {
           props: {
@@ -388,14 +399,14 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'test description'
 			})
 		})
 	})
 
 	describe('v-model', () => {
-		it('should set the @model property as v-model instead of test', () => {
+		it('should set the @model property as v-model instead of test', async () => {
 			const src = `
         export default {
           props: {
@@ -407,14 +418,14 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'test description'
 			})
 			expect(documentation.getPropDescriptor).not.toHaveBeenCalledWith('test')
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('v-model')
 		})
 
-		it('should set the @model property as v-model instead of value even with a type', () => {
+		it('should set the @model property as v-model instead of value even with a type', async () => {
 			const src = `
         export default {
           props: {
@@ -429,14 +440,14 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'Binding from v-model'
 			})
 			expect(documentation.getPropDescriptor).not.toHaveBeenCalledWith('value')
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('v-model')
 		})
 
-		it('should set the v-model instead of value with model property', () => {
+		it('should set the v-model instead of value with model property', async () => {
 			const src = `
         export default {
       model:{
@@ -453,14 +464,14 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'Value of the field'
 			})
 			expect(documentation.getPropDescriptor).not.toHaveBeenCalledWith('value')
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('v-model')
 		})
 
-		it('should not set the v-model instead of value if model property has only event', () => {
+		it('should not set the v-model instead of value if model property has only event', async () => {
 			const src = `
         export default {
       model:{
@@ -477,7 +488,7 @@ describe('propHandler', () => {
           }
         }
         `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'Value of the field'
 			})
 			expect(documentation.getPropDescriptor).not.toHaveBeenCalledWith('v-model')
@@ -486,7 +497,7 @@ describe('propHandler', () => {
 	})
 
 	describe('@values tag parsing', () => {
-		it('should parse the @values tag as its own', () => {
+		it('should parse the @values tag as its own', async () => {
 			const src = `
   export default {
     props: {
@@ -502,7 +513,7 @@ describe('propHandler', () => {
     }
   }
   `
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				description: 'color of the component',
 				values: ['dark', 'light', 'red', 'blue'],
 				tags: {
@@ -517,7 +528,7 @@ describe('propHandler', () => {
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('color')
 		})
 
-		it('should check the validator method for super standard values', () => {
+		it('should check the validator method for super standard values', async () => {
 			const src = `
   export default {
     props: {
@@ -525,32 +536,54 @@ describe('propHandler', () => {
 			type: String,
 			validator(va){
 				return ['dark', 'light', 'red', 'blue'].indexOf(va) > -1
-			} 
+			}
         }
     }
   }
   `
-			expect(parserTest(src).values).toMatchObject(['dark', 'light', 'red', 'blue'])
+			expect((await parserTest(src)).values).toMatchObject(['dark', 'light', 'red', 'blue'])
 		})
 
-		it('should check the validator arrow function for super standard values', () => {
+		it('should check the validator arrow function for inline values', async () => {
 			const src = `
   export default {
     props: {
         color: {
 			type: String,
-			validator: (va) => 
+			validator: (va) =>
 				['dark', 'light', 'red', 'blue'].indexOf(va) > -1
         }
     }
   }
   `
-			expect(parserTest(src).values).toMatchObject(['dark', 'light', 'red', 'blue'])
+			expect((await parserTest(src)).values).toMatchObject(['dark', 'light', 'red', 'blue'])
+		})
+
+		it('should check the validator method for identifiers', async () => {
+			const src = `
+  const array = ['dark', 'light', 'red', 'blue']
+  export default {
+    props: {
+        color: {
+			type: String,
+			validator(va){
+				return array.indexOf(va) > -1
+			}
+        }
+    }
+  }
+  `
+			expect((await parserTest(src, undefined, babylon().parse(src))).values).toMatchObject([
+				'dark',
+				'light',
+				'red',
+				'blue'
+			])
 		})
 	})
 
 	describe('typescript Vue.extends', () => {
-		it('should be ok with Prop', () => {
+		it('should be ok with Prop', async () => {
 			const src = `
         export default Vue.extend({
         props: {
@@ -560,7 +593,7 @@ describe('propHandler', () => {
           }
         }
         });`
-			expect(parserTest(src, ['typescript'])).toMatchObject({
+			expect(await parserTest(src, ['typescript'])).toMatchObject({
 				type: {
 					name: 'SelectOption["value"]'
 				},
@@ -569,7 +602,7 @@ describe('propHandler', () => {
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('tsvalue')
 		})
 
-		it('should parse values in TypeScript typings', () => {
+		it('should parse values in TypeScript typings', async () => {
 			const src = `
         export default Vue.extend({
         props: {
@@ -579,7 +612,7 @@ describe('propHandler', () => {
           }
         }
         });`
-			expect(parserTest(src, ['typescript'])).toMatchObject({
+			expect(await parserTest(src, ['typescript'])).toMatchObject({
 				values: ['foo', 'bar'],
 				type: {
 					name: 'string'
@@ -589,7 +622,7 @@ describe('propHandler', () => {
 			expect(documentation.getPropDescriptor).toHaveBeenCalledWith('tsvalue')
 		})
 
-		it('should understand As anotations at the end of a prop definition', () => {
+		it('should understand As anotations at the end of a prop definition', async () => {
 			const src = `
       export default Vue.extend({
         props: {
@@ -599,7 +632,7 @@ describe('propHandler', () => {
         } as PropOptions<SocialNetwork[]>,
         }
       });`
-			expect(parserTest(src, ['typescript'])).toMatchObject({
+			expect(await parserTest(src, ['typescript'])).toMatchObject({
 				type: {
 					name: 'SocialNetwork[]'
 				},
@@ -610,7 +643,7 @@ describe('propHandler', () => {
 			})
 		})
 
-		it('should understand "as const" in prop default values', () => {
+		it('should understand "as const" in prop default values', async () => {
 			const src = `
 	export default Vue.extend({
 		props: {
@@ -620,7 +653,7 @@ describe('propHandler', () => {
 			},
 		}
 	});`
-			expect(parserTest(src, ['typescript'])).toMatchObject({
+			expect(await parserTest(src, ['typescript'])).toMatchObject({
 				defaultValue: {
 					value: '"cover"'
 				}
@@ -629,7 +662,7 @@ describe('propHandler', () => {
 	})
 
 	describe('@type', () => {
-		it('should use @type typings', () => {
+		it('should use @type typings', async () => {
 			const src = `
       export default {
         props: {
@@ -642,14 +675,14 @@ describe('propHandler', () => {
         },
         }
       };`
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				type: {
 					name: '{ bar: number, foo: string }'
 				}
 			})
 		})
 
-		it('should extract values from @type typings', () => {
+		it('should extract values from @type typings', async () => {
 			const src = `
       export default {
         props: {
@@ -662,7 +695,7 @@ describe('propHandler', () => {
         },
         }
       };`
-			expect(parserTest(src)).toMatchObject({
+			expect(await parserTest(src)).toMatchObject({
 				values: ['bar + boo', 'foo & baz'],
 				type: {
 					name: 'string'