+++ packages/vue-docgen-api/tests/components/external-validator/external-props.js
@@ -1,0 +1,3 @@
+export const externalProps = ['a', 'b', 'c']
+
+export default ['d', 'e', 'f']