+++ packages/vue-styleguidist/src/bin/styleguidist.ts
@@ -1,10 +1,10 @@
-#!/usr/bin/env node
 /* eslint-disable no-console */
 
 import minimist from 'minimist'
 import kleur from 'kleur'
 import createLogger from 'glogg'
 import StyleguidistError from 'react-styleguidist/lib/scripts/utils/error'
+import { version as vueVersion } from 'vue/package.json'
 import { SanitizedStyleguidistConfig } from '../types/StyleGuide'
 import getConfig from '../scripts/config'
 import consts from '../scripts/consts'
@@ -35,9 +35,6 @@ process.on('uncaughtException', (err: any) => {
 	}
 	process.exit(1)
 })
-
-// Make sure user has webpack installed
-const vueVersion = require('vue/package.json').version as string
 
 let correctVueVersion = false
 if (vueVersion) {