+++ packages/vue-styleguidist/src/scripts/utils/mergeWebpackConfig.ts
@@ -53,9 +53,9 @@ function isFunc(
 export default function mergeWebpackConfig(
 	baseConfig: Configuration,
 	userConfig: Configuration | ((env: string) => Configuration),
-	env: string
+	env_: string
 ): Configuration {
-	const userConfigObject = isFunc(userConfig) ? userConfig(env) : userConfig
-	const safeUserConfig = omit(userConfigObject, IGNORE_SECTIONS.concat(IGNORE_SECTIONS_ENV[env]))
+	const userConfigObject = isFunc(userConfig) ? userConfig(env_) : userConfig
+	const safeUserConfig = omit(userConfigObject, IGNORE_SECTIONS.concat(IGNORE_SECTIONS_ENV[env_]))
 	return merge(baseConfig, safeUserConfig)
 }